+++ tiger-testenv-mgr/src/test/java/de/gematik/test/tiger/testenvmgr/TestEnvManagerPositive.java
@@ -18,7 +18,6 @@ import java.util.Map;
 import lombok.Getter;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.io.FileUtils;
-import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
@@ -127,12 +126,12 @@ public class TestEnvManagerPositive extends AbstractTestTigerTestEnvMgr {
         + "    type: externalJar\n"
         + "    source:\n"
         + "      - local:winstone.jar\n"
+        + "    healthcheckUrl: http://127.0.0.1:${free.port.0}\n"
         + "    externalJarOptions:\n"
         + "      workingDir: 'target/'\n"
         + "      arguments:\n"
         + "        - --httpPort=${free.port.0}\n"
-        + "        - --webroot=.\n"
-        + "      healthcheck: http://127.0.0.1:${free.port.0}\n",
+        + "        - --webroot=.\n",
         skipEnvironmentSetup = true)
     public void testCreateExternalJarRelativePathWithWorkingDir(TigerTestEnvMgr envMgr) {
         envMgr.setUpEnvironment();
@@ -144,11 +143,11 @@ public class TestEnvManagerPositive extends AbstractTestTigerTestEnvMgr {
         + "    type: externalJar\n"
         + "    source:\n"
         + "      - local:target/winstone.jar\n"
+        + "    healthcheckUrl: http://127.0.0.1:${free.port.0}\n"
         + "    externalJarOptions:\n"
         + "      arguments:\n"
         + "        - --httpPort=${free.port.0}\n"
-        + "        - --webroot=.\n"
-        + "      healthcheck: http://127.0.0.1:${free.port.0}\n",
+        + "        - --webroot=.\n",
         skipEnvironmentSetup = true)
     public void testCreateExternalJarRelativePathWithoutWorkingDir(TigerTestEnvMgr envMgr) {
         envMgr.setUpEnvironment();
@@ -164,7 +163,7 @@ public class TestEnvManagerPositive extends AbstractTestTigerTestEnvMgr {
         createTestEnvMgrSafelyAndExecute(envMgr -> {
             CfgServer srv = envMgr.getConfiguration().getServers().get("testExternalJarMVP");
             srv.getExternalJarOptions().setWorkingDir("NonExistingFolder");
-            srv.getExternalJarOptions().setHealthcheck("NONE");
+            srv.setHealthcheckUrl("NONE");
             srv.setStartupTimeoutSec(1);
             try {
                 envMgr.setUpEnvironment();
@@ -180,8 +179,8 @@ public class TestEnvManagerPositive extends AbstractTestTigerTestEnvMgr {
         "    type: externalJar\n" +
         "    source:\n" +
         "      - \"http://localhost:${mockserver.port}/download\"\n" +
+        "    healthcheckUrl: http://127.0.0.1:${free.port.0}\n" +
         "    externalJarOptions:\n" +
-        "      healthcheck: http://127.0.0.1:${free.port.0}\n" +
         "      arguments:\n" +
         "        - \"--httpPort=${free.port.0}\"\n" +
         "        - \"--webroot=.\"\n")